+++ Connecto.App/Controllers/VendorController.cs
@@ -28,7 +28,7 @@ namespace Connecto.App.Controllers
         [HttpPost]
         public JsonResult Create(Vendor item)
         {
-            var errors = new VendorModelValidator(item).Validate();
+            var errors = new VendorValidator(item, _repo).Validate();
             if (errors.Count > 0) return Json(new ConnectoValidation { Status = "Failure", Exceptions = errors }, JsonRequestBehavior.AllowGet);
 
             item.LocationId = 1;
@@ -45,6 +45,9 @@ namespace Connecto.App.Controllers
         [HttpPost]
         public ActionResult Edit(Vendor item)
         {
+            var errors = new VendorValidator(item, _repo).Validate();
+            if (errors.Count > 0) return Json(new ConnectoValidation { Status = "Failure", Exceptions = errors }, JsonRequestBehavior.AllowGet);
+
             item.EditedBy = User.UserId();
             item.EditedOn = DateTime.Now;
             _repo.Edit(item);
@@ -56,6 +59,9 @@ namespace Connecto.App.Controllers
         [HttpPost]
         public ActionResult Delete(int id)
         {
+            var errors = new VendorValidator(_repo).Validate(id);
+            if (errors.Count > 0) return Json(new ConnectoValidation { Status = "Failure", Exceptions = errors }, JsonRequestBehavior.AllowGet);
+
             _repo.Delete(id, User.UserId());
             return Json(true, JsonRequestBehavior.AllowGet);
         }
